scripts/requirements.txt


qdrant-client
sentence-transformers
datasets 
python-dotenv
tqdm 



requests
datasets
tqdm
pyarrow